LayoutTests/accessibility/placeholder.html

<html>
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">
 
    <input placeholder="search" type="text" name="q" size="15" maxlength="255" id="searchterm" />

    <input id="password" class="field" type="password" name="sc1798" value="" placeholder="Password" spellcheck="true" maxlength="5096">
    
    <p id="description"></p>
    <div id="console"></div>
     
    <script>
        if (window.accessibilityController) {
            description("This test makes sure that the placeholder is returned as the correct attribute");

            var field = document.getElementById("searchterm").focus();
            var fieldElement = accessibilityController.focusedElement;
            shouldBe("fieldElement.stringAttributeValue('AXPlaceholderValue')", "'search'");

            document.getElementById("password").focus();
            var pass = accessibilityController.focusedElement;
            shouldBe("pass.stringAttributeValue('AXPlaceholderValue')", "'Password'");

        }
    </script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
